Test.Cardano.Ledger.Conformance

Documentation

class (ShelleyEraImp era, STS (EraRule rule era), BaseM (EraRule rule era) ~ ShelleyBase, KnownSymbol rule, ForAllExecSpecRep NFData rule era, ForAllExecSpecRep ToExpr rule era, ForAllExecTypes NFData rule era, ForAllExecTypes ToExpr rule era, ForAllExecTypes EncCBOR rule era, EncCBOR (ExecContext rule era), Eq (SpecState rule era), SpecNormalize (SpecState rule era), NFData (ExecContext rule era), NFData (PredicateFailure (EraRule rule era)), NFData (SpecTRC rule era), ToExpr (ExecContext rule era), ToExpr (PredicateFailure (EraRule rule era))) ⇒ ExecSpecRule (rule ∷ Symbol) era where

Minimal complete definition

Nothing

Associated Types

type ExecContext rule era

type ExecContext rule era = ()

type SpecEnvironment rule era

type SpecEnvironment rule era = SpecRep (Environment (EraRule rule era))

type SpecState rule era

type SpecState rule era = SpecRep (State (EraRule rule era))

type SpecSignal rule era

type SpecSignal rule era = SpecRep (Signal (EraRule rule era))

Methods

runAgdaRule ∷ HasCallStack ⇒ SpecTRC rule era → Either Text (SpecState rule era)

runAgdaRuleWithDebug ∷ HasCallStack ⇒ SpecTRC rule era → Either Text (SpecState rule era, Text)

translateInputs ∷ HasCallStack ⇒ ExecContext rule era → TRC (EraRule rule era) → Either Text (SpecTRC rule era)

default translateInputs ∷ (SpecTranslate (ExecContext rule era) (Environment (EraRule rule era)), SpecTranslate (ExecContext rule era) (State (EraRule rule era)), SpecTranslate (ExecContext rule era) (Signal (EraRule rule era)), SpecRep (Environment (EraRule rule era)) ~ SpecEnvironment rule era, SpecRep (State (EraRule rule era)) ~ SpecState rule era, SpecRep (Signal (EraRule rule era)) ~ SpecSignal rule era) ⇒ ExecContext rule era → TRC (EraRule rule era) → Either Text (SpecTRC rule era)

translateOutput ∷ ExecContext rule era → TRC (EraRule rule era) → State (EraRule rule era) → Either Text (SpecState rule era)

default translateOutput ∷ (SpecTranslate (ExecContext rule era) (State (EraRule rule era)), SpecRep (State (EraRule rule era)) ~ SpecState rule era) ⇒ ExecContext rule era → TRC (EraRule rule era) → State (EraRule rule era) → Either Text (SpecState rule era)

extraInfo ∷ HasCallStack ⇒ Globals → ExecContext rule era → TRC (EraRule rule era) → Either Text (State (EraRule rule era), [Event (EraRule rule era)]) → Doc AnsiStyle

data ConformanceResult rule era

Constructors

ConformanceResult

Fields

crTranslationResult ∷ Either (NonEmpty (PredicateFailure (EraRule rule era))) (SpecState rule era)

crSpecificationResult ∷ Either Text (SpecState rule era)

crImplementationResult ∷ Either (NonEmpty (PredicateFailure (EraRule rule era))) (State (EraRule rule era), [Event (EraRule rule era)])

data SpecTRC rule era

Constructors

SpecTRC

Fields

strcEnvironment ∷ SpecEnvironment rule era

strcState ∷ SpecState rule era

strcSignal ∷ SpecSignal rule era

type ForAllExecSpecRep (c ∷ Type → Constraint) rule era = (c (SpecEnvironment rule era), c (SpecState rule era), c (SpecSignal rule era))

type ForAllExecTypes (c ∷ Type → Constraint) rule era = (c (Environment (EraRule rule era)), c (State (EraRule rule era)), c (Signal (EraRule rule era)))

class SpecNormalize a where

Minimal complete definition

Nothing

Methods

specNormalize ∷ a → a

default specNormalize ∷ (Generic a, GSpecNormalize (Rep a)) ⇒ a → a

class SpecTranslate ctx a where

Associated Types

type SpecRep a ∷ Type

Methods

toSpecRep ∷ a → SpecTransM ctx (SpecRep a)

newtype OpaqueErrorString

OpaqueErrorString behaves like unit in comparisons, but contains an error string that can be displayed.

Constructors

OpaqueErrorString (NonEmpty Text)

data SpecTransM ctx a

generatesWithin ∷ ∀ a. (NFData a, ToExpr a, Typeable a, HasCallStack) ⇒ Gen a → Int → Spec

runConformance ∷ ∀ (rule ∷ Symbol) era. ExecSpecRule rule era ⇒ ExecContext rule era → TRC (EraRule rule era) → ImpTestM era (ConformanceResult rule era)

checkConformance ∷ ∀ rule era. (HasCallStack, Era era, EncCBOR (ExecContext rule era), EncCBOR (Environment (EraRule rule era)), EncCBOR (State (EraRule rule era)), EncCBOR (Signal (EraRule rule era)), ToExpr (SpecState rule era), Eq (SpecState rule era), SpecNormalize (SpecState rule era)) ⇒ ExecContext rule era → TRC (EraRule rule era) → Either Text (SpecState rule era) → Either Text (SpecState rule era) → ImpTestM era ()

testConformance ∷ ∀ rule era. (HasCallStack, ExecSpecRule rule era) ⇒ ExecContext rule era → TRC (EraRule rule era) → Property

translateWithContext ∷ SpecTranslate ctx a ⇒ ctx → a → ImpTestM era (Either Text (SpecRep a))

Translate a Haskell type a whose translation context is ctx into its Agda type, in the ImpTest monad.

diffConformance ∷ ToExpr a ⇒ Either Text a → Either Text a → Doc AnsiStyle

runFromAgdaFunction ∷ (SpecEnvironment rule era → SpecState rule era → SpecSignal rule era → ComputationResult Text (SpecState rule era)) → SpecTRC rule era → Either Text (SpecState rule era)

runSpecTransM ∷ ctx → SpecTransM ctx a → Either Text a

unComputationResult ∷ ComputationResult Text a → Either Text a

askCtx ∷ ∀ b ctx. Inject ctx b ⇒ SpecTransM ctx b

withCtx ∷ ctx → SpecTransM ctx a → SpecTransM ctx' a

unComputationResult_ ∷ ComputationResult Void a → Either e a

toSpecRep_ ∷ SpecTranslate () a ⇒ a → SpecRep a

agdaHashToBytes ∷ Int → Integer → ByteString

integerToHash ∷ ∀ h a. HashAlgorithm h ⇒ Integer → Maybe (Hash h a)

signatureFromInteger ∷ DSIGNAlgorithm v ⇒ Integer → Maybe (SigDSIGN v)

vkeyFromInteger ∷ Integer → Maybe (VKey kd)

committeeCredentialToStrictMaybe ∷ CommitteeAuthorization → StrictMaybe (Credential 'HotCommitteeRole)

hashToInteger ∷ Hash a b → Integer

vkeyToInteger ∷ VKey kd → Integer

signatureToInteger ∷ DSIGNAlgorithm v ⇒ SigDSIGN v → Integer

agdaHashToExpr ∷ Int → Integer → Expr

unionsHSMap ∷ Ord k ⇒ [HSMap k v] → HSMap k v

unionHSMap ∷ Ord k ⇒ HSMap k v → HSMap k v → HSMap k v

mapHSMapKey ∷ (k → l) → HSMap k v → HSMap l v

bimapMHSMap ∷ Applicative m ⇒ (k → m k') → (v → m v') → HSMap k v → m (HSMap k' v')